수정

+++ src/main/java/com/spring/domain/WishlistVO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class WishlistVO {
-	private int wno;
+	private Long wno;
 	private int pid;
 	private String pname;
 	private int pprice;
